Distinguishability (#110)
- inflation method for evaluating circuits with distinguishable photons

- internal state parameters for Create and DualRail

- distinguishability notebook with HOM and Fusion

---------

+++ optyx/lo.py
@@ -95,7 +95,7 @@ photons on two output modes.
 .. image:: /_static/BS_hom_2.png
     :align: center
 
->>> np.round(diagram_qpath.to_zw().to_tensor().to_quimb()^..., 1)
+>>> float(np.round(diagram_qpath.to_zw().to_tensor().to_quimb()^..., 1))
 -0.0
 
 We can also do the same using :code:`Perceval`:
@@ -624,7 +624,6 @@ BS = BBS(0)
 
 # an alternative definition of a beam splitter
 BS_matrix_hadamard = np.sqrt(1 / 2) * np.array([[1, 1], [1, -1]])
-# BS_matrix_hadamard = np.array([[1, 1], [1, -1]])
 BS_hadamard = Box("BS_hadamard", Mode(2), Mode(2))
 BS_hadamard.to_zw = lambda: matrix_to_zw(BS_matrix_hadamard)
 BS_hadamard.conjugate = lambda: BS_hadamard
